# device-reg-service/src/main/java/ru/yandex/practicum/smarthome/dto/DeviceDto.java
package ru.yandex.practicum.smarthome.dto;

import lombok.Data;
import ru.yandex.practicum.smarthome.entity.DeviceData;

@Data
public class DeviceDto {
    private Long id;
    private Long userId;
    private String name;
    private String type;
    private String serialNumber;
    private String modelName;
    private String manufacturer;
    private DeviceConfDto configuration;
}